domain/src/test/kotlin/mega/privacy/android/domain/usecase/camerauploads/AreAllUploadTransfersPausedUseCaseTest.kt
package mega.privacy.android.domain.usecase.camerauploads

import com.google.common.truth.Truth
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.runTest
import mega.privacy.android.domain.repository.TransferRepository
import org.junit.jupiter.api.BeforeAll
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.TestInstance
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.ValueSource
import org.mockito.kotlin.mock
import org.mockito.kotlin.reset
import org.mockito.kotlin.whenever

/**
 * Test class for [AreAllUploadTransfersPausedUseCase]
 */
@ExperimentalCoroutinesApi
@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class AreAllUploadTransfersPausedUseCaseTest {

    private lateinit var underTest: AreAllUploadTransfersPausedUseCase

    private val transferRepository = mock<TransferRepository>()

    @BeforeAll
    fun setUp() {
        underTest = AreAllUploadTransfersPausedUseCase(
            transferRepository = transferRepository,
        )
    }

    @BeforeEach
    fun resetMocks() {
        reset(transferRepository)
    }

    @ParameterizedTest(name = "with {0}")
    @ValueSource(booleans = [true, false])
    fun `test that correct value is returned when invoked`(expected: Boolean) =
        runTest {
            whenever(transferRepository.areAllUploadTransfersPaused()).thenReturn(expected)
            val actual = underTest()
            Truth.assertThat(actual).isEqualTo(expected)
        }
}
